ecosense-ai/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.0
httpx==0.27.0
anthropic==0.28.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
shap==0.45.0
lifelines==0.27.8
ruptures==1.1.9
joblib==1.3.2
apscheduler==3.10.4
sqlmodel==0.0.18
python-dotenv==1.0.1
